Refactor maction output wrappers to move common code to the common wrapper.
