+++ ts/output/common/Wrappers/maction.ts
@@ -37,6 +37,8 @@ import {
 import { CommonOutputJax } from '../../common.js';
 import { MmlNode } from '../../../core/MmlTree/MmlNode.js';
 import { MmlMaction } from '../../../core/MmlTree/MmlNodes/maction.js';
+import { TextNode } from '../../../core/MmlTree/MmlNode.js';
+import { STATE } from '../../../core/MathItem.js';
 import { BBox } from '../../../util/BBox.js';
 import { split } from '../../../util/string.js';
 
@@ -222,9 +224,38 @@ export interface CommonMaction<
   readonly selected: WW;
 
   /**
+   * The prefix to use for the toggle action attribute
+   */
+  readonly prefix: string;
+
+  /**
    * Look up attribute parameters
    */
   getParameters(): void;
+
+  /**
+   * Add an event handler to the output for this maction
+   *
+   * @param {string} type The event handler type.
+   * @param {EventHandler} handler The actual event handler.
+   * @param {N=} dom The DOM node. If not provided goes over all elements of
+   *the dom tree of this wrapper.
+   */
+  setEventHandler(type: string, handler: EventHandler, dom?: N): void;
+
+  /**
+   * Public access to em method (for use in notation functions)
+   *
+   * @param {number} m   The number to convert to pixels
+   * @returns {string}    The dimension with "px" units
+   */
+  Em(m: number): string;
+
+  /**
+   * @param {number} m   The number to convert to pixels
+   * @returns {string}    The dimension with "px" units
+   */
+  Px(m: number): string;
 }
 
 /**
@@ -312,7 +343,98 @@ export function CommonMactionMixin<
      * @override
      */
     /* prettier-ignore */
-    public static actions: ActionMap<N, T, D, JX, WW, WF, WC, CC, VV, DD, FD,
+    public static actions = new Map([
+      [
+        'toggle',
+        [
+          (node, _data) => {
+            //
+            // Mark which child is selected
+            //
+            node.dom.forEach((dom) => {
+              node.adaptor.setAttribute(
+                dom,
+                `${node.prefix}toggle`,
+                node.node.attributes.get('selection') as string
+              );
+            });
+            //
+            // Cache the data needed to select another node
+            //
+            const math = node.factory.jax.math;
+            const document = node.factory.jax.document;
+            const mml = node.node as MmlMaction;
+            //
+            // Add a click handler that changes the selection and rerenders the expression
+            //
+            node.setEventHandler('click', (event: Event) => {
+              if (!math.end.node) {
+                //
+                // If the MathItem was created by hand, it might not have a node
+                // telling it where to replace the existing math, so set it.
+                //
+                math.start.node = math.end.node = math.typesetRoot;
+                math.start.n = math.end.n = 0;
+              }
+              mml.nextToggleSelection();
+              math.rerender(
+                document,
+                mml.attributes.get('data-maction-id')
+                  ? STATE.ENRICHED
+                  : STATE.RERENDER
+              );
+              event.stopPropagation();
+            });
+          },
+          {},
+        ],
+      ],
+
+      [
+        'tooltip',
+        [(_node, _data) => {}, {}] // overriden in by subclasses
+      ],
+
+      [
+        'statusline',
+        [
+          (node, data) => {
+            const tip = node.childNodes[1];
+            if (!tip) return;
+            if (tip.node.isKind('mtext')) {
+              const adaptor = node.adaptor;
+              const text = (tip.node as TextNode).getText();
+              node.dom.forEach((dom) =>
+                adaptor.setAttribute(dom, `${node.prefix}statusline`, text)
+              );
+              //
+              // Set up event handlers to change the status window
+              //
+              node.setEventHandler('mouseover', (event: Event) => {
+                if (data.status === null) {
+                  const body = adaptor.body(adaptor.document);
+                  data.status = adaptor.append(
+                    body,
+                    node.html('mjx-status', {}, [node.text(text)])
+                  );
+                }
+                event.stopPropagation();
+              });
+              node.setEventHandler('mouseout', (event: Event) => {
+                if (data.status) {
+                  adaptor.remove(data.status);
+                  data.status = null;
+                }
+                event.stopPropagation();
+              });
+            }
+          },
+          {
+            status: null, // cached status line
+          },
+        ],
+      ],
+    ]) as ActionMap<N, T, D, JX, WW, WF, WC, CC, VV, DD, FD,
       FC, CommonMaction<N, T, D, JX, WW, WF, WC, CC, VV, DD, FD, FC>
     >;
 
@@ -347,6 +469,13 @@ export function CommonMactionMixin<
       return (
         this.childNodes[i] || this.wrap((this.node as MmlMaction).selected)
       );
+    }
+
+    /**
+     * @override
+     */
+    public get prefix(): string {
+      return '';
     }
 
     /**
@@ -408,6 +537,29 @@ export function CommonMactionMixin<
      */
     public computeLineBBox(i: number) {
       return this.getChildLineBBox(this.selected, i);
+    }
+
+    /**
+     * @override
+     */
+    public setEventHandler(type: string, handler: EventHandler, dom: N = null) {
+      (dom ? [dom] : this.dom).forEach((node) =>
+        (node as any).addEventListener(type, handler)
+      );
+    }
+
+    /**
+     * @override
+     */
+    public Em(m: number): string {
+      return this.em(m);
+    }
+
+    /**
+     * @override
+     */
+    public Px(m: number): string {
+      return this.px(m);
     }
   } as any as B;
 }
