+++ ts/output/svg/Wrapper.ts
@@ -782,14 +782,6 @@ export class SvgWrapper<N, T, D> extends CommonWrapper<
   }
 
   /**
-   * @param {string} text  The text from which to create an HTML text node
-   * @returns {T}  The generated text node with the given text
-   */
-  public text(text: string): T {
-    return this.jax.text(text);
-  }
-
-  /**
    * @param {number} x   The dimension to display
    * @param {number=} n  The number of digits to display
    * @returns {string}    The dimension with the given number of digits (minus trailing zeros)
